+++ CHANGELOG.md
@@ -4,12 +4,13 @@ Release notes for RoboSpice
 Version 1.5.0 (planned)
 -------------
 
-Version 1.4.6 (Planned July 14th 2013)
+Version 1.4.6 (Planned July 30th 2013)
 -------------
 
 ** Enhancements
 * Google Http Client module now has pre-set services for Jackson,Jackson2 and Gson.
-
+* Complete refactor of the core of RoboSpice : Request processor has now been split into distinct entities. Thx to Andrew Clark. pull request #160
+* addListenerIfPending can nnow notify when no pending request has been found. Thx to Andrew Clark. pull request #160
 
 ** Bug fixes
 * Spelling correction in logs by Joel Steres